requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.5.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
lxml_html_clean==0.4.2
multidict==6.1.0
newspaper4k==0.9.3.1
openai==1.77.0
anthropic==0.40.0
propcache==0.2.0
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.0
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
Telethon==1.40.0
tenacity==8.2.3
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.2.3
yarl==1.15.2
pytest-asyncio==0.23.6
supabase==2.15.3
gotrue==2.12.0
postgrest==1.0.2
realtime==2.4.3
storage3==0.11.3
supafunc==0.9.4
